1.3 DAO 확장
