+++ src/main/java/com/nahoonzzang/tobyspring/ConnectionMaker.java
@@ -1,0 +1,8 @@
+package com.nahoonzzang.tobyspring;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public interface ConnectionMaker {
+    public Connection makeConnection() throws ClassNotFoundException, SQLException;
+}
