+++ src/main/java/com/nahoonzzang/tobyspring/DConnectionMaker.java
@@ -1,0 +1,18 @@
+package com.nahoonzzang.tobyspring;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DConnectionMaker implements ConnectionMaker {
+
+    public Connection makeConnection() throws ClassNotFoundException, SQLException {
+        Class.forName("com.mysql.jdbc.Driver");
+        Connection connection = DriverManager.getConnection(
+                "jdbc:mysql://localhost/tobistudy",
+                "root",
+                "a10234"
+        );
+        return connection;
+    }
+}
