--- src/main/java/com/nahoonzzang/tobyspring/DUserDao.java
@@ -1,11 +1,0 @@
-package com.nahoonzzang.tobyspring;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-
-public class DUserDao extends UserDao {
-    @Override
-    public Connection getConnection() {
-        return null;
-    }
-}
