+++ src/main/java/com/nahoonzzang/tobyspring/UserDao.java
@@ -2,10 +2,15 @@ package com.nahoonzzang.tobyspring;
 
 import java.sql.*;
 
-public abstract class UserDao {
+public class UserDao {
+    private ConnectionMaker connectionMaker;
+
+    public UserDao(ConnectionMaker connectionMaker) {
+        this.connectionMaker = connectionMaker;
+    }
 
     public void add(User user) throws ClassNotFoundException, SQLException {
-        Connection connection = getConnection();
+        Connection connection = connectionMaker.makeConnection();
 
         PreparedStatement preparedStatement= connection.prepareStatement(
                 "insert into users(id, name, password) values(?,?,?)");
@@ -20,8 +25,7 @@ public abstract class UserDao {
     }
 
     public User get(String id) throws ClassNotFoundException, SQLException {
-        Class.forName("com.mysql.jdbc.Driver");
-        Connection connection = getConnection();
+        Connection connection = connectionMaker.makeConnection();;
 
         PreparedStatement preparedStatement = connection.prepareStatement(
                 "select * from users where id = ?");
@@ -40,6 +44,4 @@ public abstract class UserDao {
 
         return user;
     }
-
-    public abstract Connection getConnection() throws ClassNotFoundException, SQLException;
 }
